test: fixed data urls after moving
