+++ test/integration/test_cases/web_content/removed/animated/composition_comparison_waterfall_column_2dis_1con.mjs
@@ -1,4 +1,4 @@
-import { data } from "../../../test_data/chart_types_eu.mjs";
+import { data } from "../../../../test_data/chart_types_eu.mjs";
 
 const testSteps = [
   (chart) =>
