+++ test/integration/test_cases/web_content/removed/animated/merge_split_area_stream_3dis_1con.mjs
@@ -1,4 +1,4 @@
-import { data_6 } from "../../../test_data/chart_types_eu.mjs";
+import { data_6 } from "../../../../test_data/chart_types_eu.mjs";
 
 const testSteps = [
   (chart) =>
